Описание изменений

+++ C#_MyProject/DataBaseStock.cs
@@ -11,9 +11,9 @@ namespace C__MyProject
     {
         string path = @"Stock.xml";
 
-        public void SaveAllData(SortedDictionary<int, Details> goods)
+        public void SaveAllData(SortedDictionary<long, Details> goods)
         {
-            var serializer = new XmlSerializer(typeof(SortedDictionary<int, Details>));
+            var serializer = new XmlSerializer(typeof(SortedDictionary<long, Details>));
             var fs = new FileStream(path, FileMode.Create, FileAccess.Write);
             StreamWriter sw = new StreamWriter(fs);
             serializer.Serialize(sw, goods);
@@ -22,9 +22,9 @@ namespace C__MyProject
 
         public void LoadAllData()
         {
-            var serializer = new XmlSerializer(typeof(SortedDictionary<int, Details>));
+            var serializer = new XmlSerializer(typeof(SortedDictionary<long, Details>));
             var read = new FileStream(path, FileMode.Open, FileAccess.Read);
-            var copy = (SortedDictionary<int, Details>)serializer.Deserialize(read);
+            var copy = (SortedDictionary<long, Details>)serializer.Deserialize(read);
 
             read.Close();
